new code ready to merge
