+++ jokes.json
@@ -1,45 +1,45 @@
 {
-    "jokes": [
-    {
-        "part1": "What’s the best thing about Switzerland?",
-        "part2": "I don’t know, but the flag is a big plus."
-    },
-    {
-        "part1": "Hear about the new restaurant called Karma?",   
-        "part2": "There’s no menu: You get what you deserve."
+	"jokes": [
 
-    },
-    {
-        "part1": "What is an astronaut’s favorite part on a computer?",
-        "part2": "The space bar."
-    },
-    {
-        "part1": "Why aren’t koalas actual bears?",
-        "part2": "They don’t meet the koalafications."
-    },
-    {
-        "part1": "Rest in peace to boiling water.",
-        "part2": "You will be mist."
-    },
-    {
-        "part1": "Want to hear a construction joke?",
-        "part2": "Oh never mind, I’m still working on that one."
-    },
-    {
-        "part1": "Why did the gym close down?",
-        "part2": "It just didn’t work out!"
-    },
-    {
-        "part1": "Two artists had an art contest.",
-        "part2": "It ended in a draw!"
-    },
-    {
-        "part1": "Why doesn’t the sun go to college?",
-        "part2": "Because it has a million degrees!"
-    },
-    {
-        "part1": "Why are skeletons so calm?",
-        "part2": "Because nothing gets under their skin?"
-    }
-        ]
+		{
+			"part1": "What’s the best thing about Switzerland?",
+			"part2": "I don’t know, but the flag is a big plus."
+		},
+		{
+			"part1": "Hear about the new restaurant called Karma?",
+			"part2": "There’s no menu: You get what you deserve."
+		},
+		{
+			"part1": "What is an astronaut’s favorite part on a computer?",
+			"part2": "The space bar."
+		},
+		{
+			"part1": "Why aren’t koalas actual bears?",
+			"part2": "They don’t meet the koalafications."
+		},
+		{
+			"part1": "Rest in peace to boiling water.",
+			"part2": "You will be mist."
+		},
+		{
+			"part1": "Want to hear a construction joke?",
+			"part2": "Oh never mind, I’m still working on that one."
+		},
+		{
+			"part1": "Why did the gym close down?",
+			"part2": "It just didn’t work out!"
+		},
+		{
+			"part1": "Two artists had an art contest.",
+			"part2": "It ended in a draw!"
+		},
+		{
+			"part1": "Why doesn’t the sun go to college?",
+			"part2": "Because it has a million degrees!"
+		},
+		{
+			"part1": "Why are skeletons so calm?",
+			"part2": "Because nothing gets under their skin?"
+		}
+	]
 }
